+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Tooltip: avoid removing the anchor `aria-describedby`
+
 ## [3.6.4][] - 2024-02-20
 
 ### Fixed
@@ -1898,7 +1902,5 @@ _Failed released_
 [3.6.2]: https://github.com/lumapps/design-system/tree/v3.6.2
 [unreleased]: https://github.com/lumapps/design-system/compare/v3.6.3...HEAD
 [3.6.3]: https://github.com/lumapps/design-system/tree/v3.6.3
-
-
-[Unreleased]: https://github.com/lumapps/design-system/compare/v3.6.4...HEAD
+[unreleased]: https://github.com/lumapps/design-system/compare/v3.6.4...HEAD
 [3.6.4]: https://github.com/lumapps/design-system/tree/v3.6.4